pyproject.toml
[tool.poetry]
name = "repl_python3_Tango"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
python-dotenv = "^0.18.0"
aiosqlite = "^0.17.0"
config = "^0.5.0"
Flask = "^2.0.1"
humanize = "^3.10.0"
animec = "^0.3.3"
"dislash.py" = "^1.4.2"
jishaku = "^2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
